fixed crawler bug in big novels

+++ utils/handler.py
@@ -214,12 +214,12 @@ class FileHandler:
 
     async def crawlnsend(
             self, ctx: commands.Context, bot: Raizel, title: str, title_name: str, originallanguage: str
-    ) -> None:
+    ) -> str:
         download_url = None
         if (size := os.path.getsize(f"{title}.txt")) > 8 * 10 ** 6:
-            if size > 25 * 10 ** 6:
+            if size > 25 * 10 ** 6 and int(bot.crawler[ctx.author.id].split("/")[1]) > 2000:
                 os.remove(f"{title}.txt")
-                await ctx.send('Crawled file is too big. there is some problem in crawler')
+                return await ctx.send('Crawled file is too big. there is some problem in crawler')
             try:
                 file = bot.mega.upload(f"{title}.txt")
                 await ctx.send(
